fix(taskworker) Move relocation namespace with the others
Move the relocation and relocation_control namespaces to live with the
other namespace declarations.
