--- src/sentry/relocation/tasks/__init__.py
@@ -1,4 +1,0 @@
-from sentry.taskworker.registry import TaskNamespace, taskregistry
-
-relocation_tasks: TaskNamespace = taskregistry.create_namespace("relocation")
-relocation_control_tasks: TaskNamespace = taskregistry.create_namespace("relocation.control")
