+++ src/sentry/relocation/tasks/transfer.py
@@ -18,10 +18,10 @@ from sentry.relocation.services.relocation_export.service import (
     control_relocation_export_service,
     region_relocation_export_service,
 )
-from sentry.relocation.tasks import relocation_control_tasks, relocation_tasks
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task
 from sentry.taskworker.config import TaskworkerConfig
+from sentry.taskworker.namespaces import relocation_control_tasks, relocation_tasks
 from sentry.types.region import get_local_region
 
 logger = logging.getLogger("sentry.relocation")
