+++ src/sentry/taskworker/namespaces.py
@@ -52,6 +52,8 @@ tempest_tasks = taskregistry.create_namespace("tempest")
 
 uptime_tasks = taskregistry.create_namespace("uptime")
 
+relocation_tasks = taskregistry.create_namespace("relocation")
+relocation_control_tasks = taskregistry.create_namespace("relocation.control")
 
 # Namespaces for testing taskworker tasks
 exampletasks = taskregistry.create_namespace(name="examples")
